BAEL-951 adding @Repository stereotype
